+++ CPAC/registration/registration.py
@@ -3124,6 +3124,39 @@ def overwrite_transform_anat_to_template(wf, cfg, strat_pool, pipe_num, opt=None
 
 
 @nodeblock(
+    name="mask_sbref",
+    switch=[
+        ["registration_workflows", "functional_registration", "coregistration", "run"],
+        [
+            "registration_workflows",
+            "functional_registration",
+            "coregistration",
+            "func_input_prep",
+            "mask_sbref",
+        ],
+    ],
+    inputs=["sbref", "space-bold_desc-brain_mask"],
+    outputs=["sbref"],
+)
+def mask_sbref(wf, cfg, strat_pool, pipe_num, opt=None):
+    """Mask sbref with brain mask."""
+    mask_sbref = pe.Node(interface=afni.Calc(), name=f"mask_sbref_{pipe_num}")
+
+    mask_sbref.inputs.expr = "a*b"
+    mask_sbref.inputs.outputtype = "NIFTI_GZ"
+
+    node, out = strat_pool.get_data("sbref")
+    wf.connect(node, out, mask_sbref, "in_file_a")
+
+    node, out = strat_pool.get_data("space-bold_desc-brain_mask")
+    wf.connect(node, out, mask_sbref, "in_file_b")
+
+    outputs = {"sbref": (mask_sbref, "out_file")}
+
+    return (wf, outputs)
+
+
+@nodeblock(
     name="coregistration_prep_vol",
     switch=["functional_preproc", "run"],
     option_key=[
@@ -3134,7 +3167,7 @@ def overwrite_transform_anat_to_template(wf, cfg, strat_pool, pipe_num, opt=None
         "input",
     ],
     option_val="Selected_Functional_Volume",
-    inputs=[("desc-brain_bold", ["desc-motion_bold", "bold"], "sbref")],
+    inputs=[("desc-preproc_bold", "sbref")],
     outputs=["sbref"],
 )
 def coregistration_prep_vol(wf, cfg, strat_pool, pipe_num, opt=None):
@@ -3149,14 +3182,7 @@ def coregistration_prep_vol(wf, cfg, strat_pool, pipe_num, opt=None):
         outputtype="NIFTI_GZ",
     )
 
-    if not cfg.registration_workflows["functional_registration"]["coregistration"][
-        "func_input_prep"
-    ]["reg_with_skull"]:
-        node, out = strat_pool.get_data("desc-brain_bold")
-    else:
-        # TODO check which file is functional_skull_leaf
-        # TODO add a function to choose brain or skull?
-        node, out = strat_pool.get_data(["desc-motion_bold", "bold"])
+    node, out = strat_pool.get_data("desc-preproc_bold")
 
     wf.connect(node, out, get_func_volume, "in_file_a")